feat: migrate rest_api/test_views
